+++ packages/react-ui/src/theme/sections/palette.js
@@ -1,4 +1,4 @@
-import { alpha, darken, lighten } from '@material-ui/core';
+import { alpha, darken, lighten } from '@mui/material';
 
 const COLOR_BLACK = '#2C3032';
 const COLOR_WHITE = '#FFFFFF';
@@ -175,6 +175,7 @@ export const commonPalette = {
     light: baseColors.neutral[100],
     main: baseColors.neutral[200],
     dark: baseColors.neutral[50],
+    // are we missing here relatedLight ??
     contrastText: baseColors.common.black,
     background: baseColors.shades.dark[8]
   },