+++ packages/react-ui/src/theme/themeUtils.js
@@ -1,4 +1,12 @@
-import createSpacing from '@material-ui/core/styles/createSpacing';
+import { createSpacing } from '@mui/system';
 import { SPACING } from './themeConstants';
 
 export const spacing = createSpacing(SPACING);
+
+// Convert pixels to rem
+export function getPixelToRem(px) {
+  const fontBase = 16;
+  const rem = (1 / fontBase) * px + 'rem';
+
+  return rem;
+}