+++ packages/react-ui/src/widgets/WrapperWidgetUI.js
@@ -1,6 +1,6 @@
 import React, { useState, createRef } from 'react';
 import PropTypes from 'prop-types';
-import { makeStyles } from '@material-ui/core/styles';
+import makeStyles from '@mui/styles/makeStyles';
 import {
   Box,
   Button,
@@ -13,8 +13,8 @@ import {
   MenuItem,
   Tooltip,
   Typography
-} from '@material-ui/core';
-import { ExpandLess, ExpandMore, MoreVert } from '@material-ui/icons';
+} from '@mui/material';
+import { ExpandLess, ExpandMore, MoreVert } from '@mui/icons-material';
 
 /*
 Options props must have this format:
@@ -157,6 +157,7 @@ function WrapperWidgetUI(props) {
         aria-label={action.label}
         onClick={action.action}
         className={classes.iconAction}
+        size='large'
       >
         {action.icon}
       </IconButton>
@@ -217,6 +218,7 @@ function WrapperWidgetUI(props) {
                 aria-haspopup='true'
                 onClick={handleClick}
                 className={classes.iconAction}
+                size='large'
               >
                 {optionsIcon}
               </IconButton>