+++ packages/react-ui/storybook/.storybook/preview.js
@@ -1,35 +1,28 @@
 import React from 'react';
 import { withDesign } from 'storybook-addon-designs';
-import { createTheme, responsiveFontSizes, ThemeProvider } from '@material-ui/core';
+import {
+  createTheme,
+  responsiveFontSizes,
+  ThemeProvider,
+  StyledEngineProvider,
+  adaptV4Theme
+} from '@mui/material';
 import { cartoThemeOptions } from '../../src/theme/carto-theme';
 
-let theme = createTheme(cartoThemeOptions);
+let theme = createTheme(adaptV4Theme(cartoThemeOptions));
 theme = responsiveFontSizes(theme, {
-  breakpoints: cartoThemeOptions.breakpoints.keys,
+  breakpoints: ['sm'],
   disableAlign: false,
-  factor: 2,
-  variants: [
-    'h1',
-    'h2',
-    'h3',
-    'h4',
-    'h5',
-    'h6',
-    'subtitle1',
-    'subtitle2',
-    'body1',
-    'body2',
-    'button',
-    'caption',
-    'overline'
-  ]
+  factor: 2
 });
 
 export const decorators = [
   (Story) => (
-    <ThemeProvider theme={theme}>
-      <Story />
-    </ThemeProvider>
+    <StyledEngineProvider injectFirst>
+      <ThemeProvider theme={theme}>
+        <Story />
+      </ThemeProvider>
+    </StyledEngineProvider>
   )
 ];