+++ packages/react-ui/storybook/stories/common/Dialog.stories.js
@@ -10,7 +10,7 @@ import {
   Grid,
   Slide,
   TextField
-} from '@material-ui/core';
+} from '@mui/material';
 
 const options = {
   title: 'Common/Dialog',